requirements.txt
matplotlib==3.8.3 
numpy==1.26.4 
pandas==2.2.1 
tensorflow==2.16.1 
pillow==10.2.0 
requests==2.31.0 
deep-translator==1.11.4 
google-generativeai==0.3.2 
python-docx==1.1.0 
python-dotenv==1.0.1 
openai==1.12.0 
streamlit==1.32.2
